Overhaul: Individual Spacing Values
The routeData now has a slightly different data structure to it

The timelines are now strictly enforced to all have the same length

Various functions have been removed and added for streamlining access to
RouteData via PhantomRoute

Recording now happens every single time recordPeriodic() is called. We
can change this if we want to.

Updated getRawAxis and getRawButton to work with the new system.

Improved printouts

WHY
Because previously if you set the spacing to 100ms, then it would record
at intervals between 100ms and 120ms. These would then be played back at
a standard spacing of 100ms, meaning the whole thing would actually be
sped up. It was weird.